+++ src/sprites/TilesetGenerator.ts
@@ -1,6 +1,0 @@
-// No longer used
-export const TILE = {} as any;
-export const TOTAL_TILES = 0;
-export class TilesetGenerator {
-  generate(): void {}
-}